Bug#28636768: ANALYZE AND BACKPORT BUG#28365855 TO 5.7
When stopping, channels that had pending transactions could deadlock
in the GR before commit hook invocation.
The stop process takes a write lock for the master info channel, but
it then waits on the complete of its transactions that might try to
take a read lock on the same structure.

As a solution we add context info the THD object.
That way the before commit hook no longer depends on channel locks to
complete its process.

It also includes some related locking corrections and improvements
with updated documented information on subject.