Bug#28746409 RESTART TIME LIMIT CALCULATION ERROR IN NDB_WAIT_UNTIL_READY.TEST
The test ndb_wait_until_ready fails on platforms where first mysqld
restart takes longer than 30 seconds(the value of --ndb-wait-connected).

This was noticed only in 8.0 based versions because mysqld restart times
is normally longer.

Fix calculation of upper limit to be first measured restart time
plus half of --ndb-wait-connected (which would be the max time that
mysqld will wait when restarting). The assumption is that second restart
will take same time as first, with some added safety.

Change order of the printouts which occur before faiulre to make it
easier to directly see in the log file what the measured and limit
values are if test fails.

Also remove line which casues the test to be skipped on valgrind, should
be fixed now.

Change-Id: Ia210ddb01ab2f076bd8c5241195f7b97e34da71d
(cherry picked from commit 329c5833c58ac516b66216180ac6cce43da89d99)