Bug #28556437 ASSERTION FAILURE: LOB0LOB.CC.*EXTERN_LEN > 0
Problem:
The failure is because of a zero size lob stored externally. This results
in tripping an assertion as the code does not expect a zero sized lob.

Solution:
Treat the lob as an empty lob if it's size happens to be 0. Also put a debug
assert that the page_no should be FIL_NULL when the lob is empty.

RB# 20667
Reviewed by: Annamalai Gurusam (annamalai.gurusami@oracle.com)