Bug#28569645: SQL INJECTION ON SLAVE DUE TO NON-QUOTING IN               BINLOGGED ROLLBACK TO SAVEPOINT
Description: SQL injection is possible on slave due to
incorrect binlogging of 'ROLLBACK TO SAVEPOINT' statements.

Analysis: Whenever an identifier is written to binary log,
it should be properly quoted to prevent any SQL injection
on the slave. For binlogging of 'ROLLBACK TO SAVEPOINT'
statements, this is correctly fixed on 5.5 by the patch
for Bug#14548159, but the fix is missing from 5.6+.

Fix: Added proper binlogging for 'ROLLBACK TO SAVEPOINT'
statements.