BUG#27442477 ASSERT `THD->GET_HA_DATA...HAS_STATE(XID_STATE::XA_ACTIVE))' AT HANDLER.CC:1396
[post-push fix]

Problem
-------
After BUG#27442477, even if an XA transaction gets emptied by filters the slave
applier will continue to run.
While applying on the slave, if a XA transaction becomes empty, the slave
applier should stop and the user should be called upon to solve the matter.

Analysis & Fix
--------------
A client error should be thrown in XA PREPARE, if the transaction is empty.
Nothing should be written to the binary log.