28726286 [INNODB] SEMAPHORE WAIT HAS LASTED > 903 SECONDS during RQG test
1) All undo tablespaces became inactive. The code in
innodb_alter_undo_tablespace_inactive() that is supposed to keep at least
2 undo spaces active at all times did not work when there are multiple
undo spaces marked inactive implicit.  It assumed there was only one.
This change takes away that assumption and only counts the
innactive_implicit space that is marked.

2) In this hang, half of the undo spaces were marked inactive_implicit.
The purge thread only marks one of those at a time.  The others are
explicitly SET INACTIVE and then SET ACTIVE again before they became empty.
This changes the alter_active() function to check if the inactive_explicit
space has been marked. If so, it is set to inactive_implicit. If not
marked yet, it is set directly to active.

Approved by Rahul in RB#20682