BUG#28747936: CONSISTENT TRANSACTION OPTIONS ARE BEING ALLOWED ON A NOT ONLINE MEMBER
The begin hook is only invoked once per transaction, to ensure that
it sets a flag on the first time that it is used.
That flag was incorrectly being set when the hook was throwing
errors, this was allowing the hook to be bypassed after the first
error.

To fix the above issue, the flag is only set when the hook is run
without errors.