Bug#28315651 INFORMATION_SCHEMA.INNODB_FOREIGN DOES NOT RETURN A CORRECT TYPE
The field "TYPE" of information_schema.innodb_foreign is always "0", but
in fact, it should be set as https://dev.mysql.com/doc/refman/8.0/en/innodb-foreign-table.html.

The root cause is we didn't compose a correct sql for information_schema.innodb_foreign.
So, in this patch, we will compose a correct SQL to generate "TYPE" for this.

Reviewed-by: Mayank Prasad <mayank.prasad@oracle.com>