Bug#28785835 WORKAROUND ASAN BUG FOR TIRPC
Sun RPC, and XDR, is being removed from glibc, and into a separate libtirpc
library. This is not compatible with libasan. The interceptor functions
inserted into the code will segfault.
As a workaround, do LD_PRELOAD=/lib64/libtirpc.so
For dynamically linked libasan (default for gcc), we must preload that as well.

This currently affects fedora, but will likely also affect other linux
variants in the future.

This patch also enables ASAN suppressions in asan.supp
This patch also enables LSAN suppressions in lsan.supp

Change-Id: I1bc777482535ce21595a48ae12679c325667d722