Bug#28581468 - RENAMING PARENT COLUMN IN A FOREIGN KEY FAILS WITH STRANGE ERROR.
The error reported during the ALTER TABLE is because the the foriegn key
cache is not updated or invalidated properly before opening the table,
so opening table will find mismatch information between the metadata
from DD and the column name in foriegn key cache.

This is finally proved to be a regression from new DD work, which mistakenly
removed the foreign key cache invalidation for the rebuild ALTER TABLE.

So the fix is just to invalidate the foreign key cache if the foreign key
is affected by the column rename. And new foreign key cache would be
re-created on next opening table.

RB: 20711
Reviewed-by: Allen Lai <zheng.lai@oracle.com>