Bug#28786951: SET LAST_QUERY_COST FOR QUERIES WITH SUBQUERIES AND UNIONS
The Last_query_cost status variable only showed the cost of flat
queries. Its value was zero for queries which contained subqueries or
unions.

This patch changes the calculation of Last_query_cost so that it also
shows the cost of queries that have multiple query blocks. It sums up
the cost estimates of each query block. Additionally, it estimates how
many times non-cacheable subqueries will be executed, and multiplies
the cost of those query blocks with the number of subquery executions.

JOIN::compare_costs_of_subquery_strategies() already had logic for
estimating the number of subquery executions. This logic is now
factored out in a separate function which is reused in the code that
calculates the last query cost.

Change-Id: Ib0d8dcc2bd0f33f73b32e314459b67d3032bb213