Bug#28727717: Duplicated 'is not null' predicates.
The optimizer may add multiple identical 'is not null' predicates
to the same where-condition.

This patch fix that be letting add_not_null_conds() check if
a particular 'is not null' predicate already existed in a 'cond'
before adding a new one.

Added specific testcase and updated several result files which
had this problem.