Post-push fix for BUG#28776365 MISSING NDB VERSION COMPATIBILITY LOGIC
The previous patch made the mistake of assuming, for example, that the
upgrade logic in 8.0.15 could determine whether 8.0.13 was compatible
with 8.0.14. In fact the logic only works when the target version
exactly matches the leftmost element in the table.

An NDB node can determine whether another node is compatible with its
own compiled-in version number, but not whether any two arbitrary
past versions are compatible with each other.  See also bug#18886034