Bug#28728603 ENABLE CONDITION PUSHDOWN FOR EQ_REF'S IN PUSHED JOINS
Rework the previous patch:

Bug#27397802 ENTIRE CLUSTER CRASHES WITH "MESSAGE TOO BIG IN SENDSIGNAL" WITH BIG IN CLAUSE

Patch for Bug#27397802 took a rather brute force approach by simply
not pushing any conditions for a lookup operations in a pushed join.

Such a lookup operation can be the parent of a branch of
pushed child operations, which will be eliminated if the condition fails.
Thus we want to reintroduce pushing of conditions for these.

This patch re-enable the usage of such pushed conditions if the
size of the generated interpreter code is less than 64 words.

This limit is choosen as a best effort value in the belief that
most conditions will be well below this limit. We also want to
be carefull in not pushing large conditions in order to reduce communication
overhead.