Bug#28745859 CONDITION PUSHDOWN SHOULD NOT BE HANDLED BY THE ITERATORS
Refactoring:

Condition push down in the iterators are required only by single table
update and delete queries. For all other types of queries it is handled
as part of the optimize phase. For all other query type the condition
push down handling in the iterators is obsolete, and may even
cause an extra round of pushability analysis which only adds overhead.

This patch introduce a call to the handler::cond_push() method in
Sql_cmd_delete::delete_from_single_table() and
Sql_cmd_update::update_from_single_table(). This makes the
condition push down handling in the iterators obsolete -> removed.