Bug #28799559: TESTS DON'T WORK WITH DEFAULT SETTING OF MAX OPEN FILE                DESCRIPTORS
Problem: Default value for variable innodb_open_files is different on different
platforms which causes result file mismatch. max_open_files vlaue is calculated
on runtime based on parameters like max_connections, table_open_cache. So if
max_open_files limit is set to a value more than the calculated value, mysqld
reports a warning.

Fix: Fix is to add a suppresion for the warning, and removed the
innodb_open_files variable as this is not needed in the test file.