Bug #28728007 HA_NDBCLUSTER::COND_PUSH() SHOULD SUPPORT PARTIAL PUSH WITH REMAINDER
Currently the pushability of a condition to NDB is limited by
that all AND'ed predicates in the condition has to be pushable
to NDB. In some cases this severely restrict the pushability
of conditions.

This patch breaks up the condition in its AND'ed predicate
components, and evaluate the pushability for each component.
If some of the predicates can not be pushed, they are
returned as a remainder condition which the server
will evaluate for us.

No separate test case is developed as the existing
test in ndb_condition_pushdown.test is sufficient
to demonstrate the effect of the patch.