Bug #28470805 DELETE CASCADE CRASHES AS FOREIGN TABLE'S VIRT COL DATA NOT UPDATED ON RESTART
PROBLEM
=======
1) We have a parent and child table & child table has a virtual column
with a secondary index defined on it.It also has delete cascade defined
on the fk constarint.
2) We try to delete a tuple from parent table after restart.
3) When doing a delete a parent row we try to load the child table in
dictionary for enforcing the fk constraint. While loading the child
table we are initiliazing the virtual column template meta data for the
virtual column.
4) Failing to initliaze it causes a crash when we are trying remove the
secondary index rows defined on virtual column .

FIX
===
In case of delete cascade, initiliaze the virtual column template meta
data info for the child table if the virtual column is affected by
foreign key.

Reviewed by : bin su <bin.x.su@oracle.com>
	      Aditya A <aditya.a@oracle.com>

RB:20686