WL#12508 PREVENT ATRT SETUP RUNNING DURING PROCESS RESTART
Adds logic to prevent running atrt setup script when restarting a process.

Modifies the atrt's start_process logic such that
 * start_process defines a new optional boolean argument called "run_setup" with
default true.
 * When the process is started for the first time, atrt-setup.sh is called
synchronizing file system configuration files (default behavior).
 *When the process is restarted, the start_process gets called from do_start_process
 with run_setup set to false, skipping atrt setup script ensuring that filesystem
is kept untouched.

In downgrade tests, cluster processes running one version are stopped and then
started with an older version. This is achieved by ATRT that stops the
process, switch its configuration, and then starts it.

The start functionality is common to both steps of starting a process for the
first time and to re-start a previously stopped process.
However, these steps have different requirements. While when starting the
process for the first time, we need to ensure a clean folder with generated
configuration files only, where  this is not the case when restarting a process.

Hence modifies logic to acheive the requirement to leave filesystem to be exactly
same as earlier as it was( with logs and configuration files) during process restart.