WL#10162 Schema dist operation error code return
Change the code which unpacks the "query" blob column to move the blob
contents directly from NdbApi into the string which will be returned.
Return the blob as a pointer to a string allocated in current memory
root. Normally there is already space available and thus this change
does not cause any additional memory allocation to occur.

Previously the code looped tiwce over all fields in the table, allocated
memory with my_malloc(), unpacked the blob into the allocated memory,
setup the field to point at the allocated memory and finally unpacked the
field into a string, then released the memory and returned the blob
content as a string.

Change-Id: Ie0e3c595ad15e93c14b69110ac6eb262f58896fa