WL#12509 CPCD: IMPROVE PROCESS'S LIFE-CYCLE LOGGING
Improves the logging of the status of the processes started by CPCD.
CPCD's log messages should involve process group, name, process's
internal CPCD id, and operating system pid to follow on process
lifecycle. And log messages are printed before the status is modified.

Existent debug log messages are updated to show process's group, name,
internal CPCD id, and it's operating system pid.
Process::removePid() calls are modified to make sure that m_pid
is set to bad_pid only after removing the pid. And the warning
log that the PID read using readPid() is valid is removed from
the readPid() function and added in the only place it is
required. The main reason for this change is that the readPid()
is not only used to fetch the pid , but also for checking whether
the process is valid or not.