Bug#27072756 NDB BINLOG: CREATING NDBEVENTOPERATION BLOB FIELD ... HANDLES FAILED (CODE=723)
When tables with BLOBs are dropped and then redefined with different number of BLOBs the
event definitions for monitoring table changes can become inconsistent. Normally when a
table is dropped the corresponding events are also dropped, but in certain error situations
(communication errors) the clean-up is not done. When tables are then re-created and conflicting
events are found they are reused, however if more BLOB columns are created some events could be
missing. To handle this we need to check the cases where:
1) We have no BLOB events, then we create them
2) We have exactly the correct number of BLOB events, reuse them.
3  We drop all BLOB events and create new ones.