Bug #28556264 - HANDLER::CREATE IS CALLED WITH ERROR ON THE CONDITION STACK
Problem:
During discover of a table from storage engine, the handler::create()
function will be called with a "Table x doen't exist" in the condition list
of the diagnostic area.
This means that a potential error which occurs in handler::create() can not
be reported properly since there is already an error on the stack.

The error is pushed early during the statement when
get_table_share_with_discover() checks if the table is installed in the DD by
calling get_table_share(). The function clears the error in diagnostic area
but leaves the warning stack untouched. Thus the warning stack contains one
error.

When get_table_share_with_discover() later check that table exists in storage
engine, error 7 will be returned and thus
handler::create(HA_OPTION_CREATE_FROM_ENGINE) will be called. This causes
handler::create() to be called with errors already pushed on stack.

Solution:
Clearing the error stack as part of get_table_share_with_discover().
We check that the function is not called with an error already set,
or with error in the condition list. If so we clear the error in the
diagnostics area and on the warning stack.

Change-Id: I409f58b9af8152d8664ceb30375bc789c3ffc7bd