BUG#28761424 - NODES ARE UNABLE TO JOIN WHEN SEEDS & LOCAL_ADDRESS IS CONFIGURED WITH HOSTNAMES
Problem
=======================

Nodes are not able to join , when group_seeds and local_address is
configured with hostnames.

Eventhough hostname is configured everywhere why are we trying  the
connection attempt with ipv6, when hostname is configured,
[Repl] Plugin group_replication reported: '[GCS] Connection attempt
from IP address XXXXXXX refused. Address is not in the IP whitelist.'
?

With default ip_whitelist, AUTOMATIC, i can see clearly all the ipv4
addresses getting whitelisted.

Analysis
=======================

In WL#11926, one decided that, when resolving a name, the preferable
address would be the V4 address. But getaddrinfo does not guarantee
order in different implementations, depending of the OS.

Solution
=======================

Since we now use getaddrinfo for setup every connections, it is a matter
of:
- Checking that we have a V4 endpoint.
- Select the 1st V4 entry and use it to connect.