BUG#27928837 `HEAD->VARIABLES.GTID_NEXT.TYPE != UNDEFINED_GTID'
[post-push]

Problem
-------
The test script `binlog_gtid.binlog_gtid_unknown_xid` is failling sporadically
on PB2 due to a difference in expect error codes:

     mysqltest: At line 110: Query 'XA ROLLBACK 'xa1'' failed with wrong error
     1397: 'XAER_NOTA: Unknown XID', should have failed with error '1399'.

Analysis
--------
This problem happens when the test is ran several times in a row, leading to the
conclusion that something has not been cleaned up properly.

After analysis, the conclusion is that an XA transaction that is started in the
test script is never commited or rolled back.

Fix
---
1) Rollback the XA transaction at the end of the test script
2) Verify that are no pending XA transactions