Bug#28818191: FIX TIME ZONE ISSUES IN NDB.CLUSTERJ TEST IN WINDOWS
The previous patch to fix the timezone issues, as a part of fix for
Bug#28424366, did not fix the issues in Windows. This patch changes the
logic with which the timezone offset is calculated to set to the mysql
session. Additionally, this patch also sets the time zone of the JVM
running the clusterj test jar to the same time zone as that of the
server.

Change-Id: Ifd4e55756a93eeef071441588005e85b3a60df63