Bug#28835066: TABLE NOT UNLOADED WITH SECONDARY_ENGINE = NULL
When SECONDARY_ENGINE is set to NULL, any existing secondary engine
definition is removed from the new table definition. Because the new
table definition was used to find the secondary engine definition, no
secondary engine was found and no unload would take place.

Use old table definition instead of the new one to determine if the
table has a secondary engine from which it should be unloaded.

Change-Id: I25a4bb5ea5ad06ef63bdb0385029ee032009896c