BUG#28658228 - [GCS] ERROR PUSHING MESSAGE INTO GROUP COMMUNICATION ENGINE
Problem
=======
While stopping GR on a server, it still tries to send messages to the group,
even after shutting down XCom. Therefore, these messages aren't sent since it
is not possible to push them into the group communication engine, and errors
are logged stating this.

Fix
===
Disregard these errors when the server is leaving or not a member of a group.