WL#12360: Add an option to allow server quick settings validation.
Prior to MySQL 8.0, 'mysqld --verbose --help' was used to verify
whether the server configuration settings was valid. Running
'mysqld --verbose --help' would result in an error if the
configuration was invalid. But with MySQL 8.0, there is no error
reported.

This WL achieves the following:

a) A new server command line/config file option called
   "validate-config" is introduced to validate the server
   configuration settings provided during server start up.

        Name: "--validate-config"
        Value: Boolean ( 0 or 1)
        Default: 0

b) "validate-config" will be a special option with the
    sole purpose of validating the options supplied. Hence
    server start up, storage engines/plugins initialization
    will not be performed when this option is used. The
    "validate-config" option will validate only the server
    options(i.e the output of "--help --verbose") and does
    not validate plugin options.

c) The server will exit with an error(error no 1) for
   first occurrence of invalid configuration.

d) The server will exit with zero value(Success) when it
   has successfully validated all the configuration options.

e) The server will exit with zero value while reporting
   warnings when invalid values are specified for valid server
   options when used with appropriate "log_error_verbosity"
   option value(except for storage engine and performance_schema
   options since they are not initialized).

f) Since the plugin initialization is not performed, when the
   plugin variables are used with "validate-config", the plugin
   variables will not be recognized and will be marked as unknown
   variables.