Bug #23329861 - WARNING ABOUT LOCALHOST WHEN USING SKIP-NAME-RESOLVE
On Unix, MySQL programs treat the host name localhost specially,
in a way that is likely different from what you expect compared
to other network-based programs. If the host is not specified or
is localhost, a connection to the local host is assumed:

On Windows, the client connects using a shared-memory connection,
if the server has shared-memory connections enabled.

On Unix, the client connects using a Unix socket file.
The --socket option or the MYSQL_UNIX_PORT environment variable
may be used to specify the socket name.

Problem:
While having --skip-name-resolve enabled one does still get
warnings and such it pollutes the logs. These warnings state
that the "localhost" entries in the user, db, and proxies_priv
will be ignored but they are used and not ignored.

Solution:
This is a hopefully temporary solution to not print warnings if
the host is localhost. Added checks to not print the warnings.
For the proper solution an ALC flag needs to be implemented.
Filed new bug for this Bug #28817102.

Change-Id: I157a5e26461b661add428b27001b3f5d5429eaa2