Bug#27586027: ASSERTION `M_STATUS == DA_EMPTY' FAILED WITH COM_STATISTICS
If the client connection is interrupted, there will be an error set on
the thd, this way, the Diagnostics_area will also have a status. Later
on, when trying to call disable_status, an assertion is triggered since
it is expected for the status to be empty when calling the function.
The fix is to check if there's a previous error before trying to disable
the status.

Change-Id: I71cac2bec394b6d81dba08fbca4a9a7be265159c