Bug#28493849: CRASH: MEMBER ACCESS WITHIN NULL POINTER OF TYPE                'STRUCT FIELD' IN
Problem:
Item_func_isnull() evaluates the result of the function while
resolving as it sees that the argument is not nullable. This
also results in Item_func_isnull() being marked as a const item.
While creating temp table, temp table field is not created
for Item_func_isnull as it is a const item. However rollup
processing presumes that all the rollup fields and expressions
having rollup fields will have a temp table field. So, while saving
data into the temp table field, server exits.

Solution:
While resolving rollup fields, maybe_null for Item_func_isnull is
marked as true. However used_tables are not updated for the Item.
This results in inconsistency. So update the used tables for all
the items after maybe_null for the item gets updated. This makes
Item_func_isnull() a non-const item for this case which enables
rollup to write the "null" correctly.