BUG#28807260: VERIFY COORDINATED ACTION SAFETY AGAINS PLUGIN STARTS AND STOPS
When most group coordinated actions execute such as
group_replication_switch_to_multi_primary_mode or
group_replication_set_as_primary they do check if the member is running
and in a majority.

The code however assumes the plugin stays running for the rest of the
checks, that may not be true when stopping.

When executing STOP GROUP_REPLICATION signal coordinator that plugin is
stopping and wait UDF terminate.