Bug #20116393 MYSQLD GOT SIGNAL 11 AFTER BEING DISCONNECTED FROM CLUSTER
Asynchronous cluster disconnection caused subsequent NdbApi
transaction start attempts to fail.

If this occurred during a 'bulk' delete, then the SQL layer
calls HA::end_bulk_delete().

The ha_ndbcluster implementation assumes that a transaction has
been started, and can crash if this is not the case.

This problem is fixed by checking that the transaction pointer
is set before using it.

A new error insertion is created and a testcase using it is added.