Bug#28448853 - ASSERTION FAILURE: ROW0LOG.CC:2153 ROW_LOG_TABLE_APPLY (THR=0X7F7C7C050030
This is due to DML(update) log written during online DDL was not logged
correctly or as expected. If there are no changes on virutal column or
the changed virtual columns are not on any index, then there is actually
no need to write any log for them. In the rollback case, the code path
will write empty record for both old and new virtual columns, which
introduces a new scenario for log parser. However, the parser only
expects a 2 bytes length marker if there is nothing to be logged, rather
than two consecutive 2 bytes length marker.

So the fix should be that if there is nothing to be logged for new
virtual column value, then there is also nothing to be logged for old
virtual column value in the update case. So only one 2 bytes length
marker would be kept for parser to understand the scenario, rather than
two markers.

RB: 20787
Reviewed-by: Sunny Bains <sunny.bains@oracle.com>