Bug#28760011: UPGRADE TO 8.0.12 FAILS
In 5.7, the table triggers were defined in files, and were expected to be
listed in a specific order. When loading the triggers from file into main
memory, the order would be corrected if it was wrong.

In 8.0, when upgrading from 5.7, we expected the trigger order in the
file to be correct, and there was no effort to correct the order.
Instead, an assert was fired if the order was wrong.

This patch implements a mechanism to correct the trigger order when
upgrading from 5.7, using the same trigger chain implementation as
was done in 5.7.

Unrelated to the above, this patch also adds a test for Bug#27960500
to verify that upgrades including FTS indexes from 8.0.11 works as
expected.

Change-Id: I5d8c1dfc9ec237d1bdf65dd0aa13f2d8816075db