BUG#28868412: ------------- execTUP_COMMITREQ should return signal->theData[0] != 0 when it has to wait for disk data page or log buffer space to become available. It doesn't explicitly set signal->theData[0] to anything in this case, fixed by setting signal->theData[0] = 1 before returning from execTUP_COMMITREQ in those cases.
BUG#27584165
------------
When scanning a row using TUP scan, ACC scan or performing a
PK read using the primary key it is possible to start a read
of the row and hit a real-time break where one has to wait
for the page to become available in memory. When the page
request returns later the row is read and crashes on wrong
checksum. The problem is that when the row is deleted its
checksum is invalidated.

Solved by introducing a new flag DELETE_WAIT on tuple header.
This flag is checked before starting any scan operation on a
row and before starting any PK read operations on the row
where disk data pages are necessary to open up.

The flag is cleared when the row is finally committed.