Bug#28875123: ASSERTION FAILURE WHILE RUNNING ALTER UNDO COMMAND AND DML IN PARALLEL
An inactive undo space can be put back to active if it has not yet been
marked for truncate. This was a non-atomic decision and action by a user
thread.  But the mark-for-truncate action can occur simultaneously by the
purge thread.  In this patch, both undo::Tablespace::alter_active() and
undo::Truncate::mark() atomic are protected by undo::Tablespace::m_rsegs::
x_lock() so that there is no race condition between a user thread doing
SET ACTIVE and the purge thread marking an undo space for truncate.