Bug#28431783 ASSERTION FALSE IN ITEM::TYPE_FOR_VARIABLE() after wf intro
Symptom:

In debug mode, we see an assert in Item::type_for_variable when that
function is called from Item_func_set_user_var::resolve_type for
args[0]->data_type() of PTI_variable_aux_set_var.

This is because args[0] is an Item_ref, which hasn't yet been fixed,
so it has an illegal type (244), as assigned by default in
Item::Item().  The case statement in Item::type_for_variable gives an
ASSERT error due to the illegal type.

This Item_ref is created in Item_field::fix_fields when we resolve the
reference "a" in the repro query[1] to an Item_outer_ref, not a field
(resolution == RESOLVED_AGAINST_ALIAS).

Before the introduction of window function (973df4495b), at this place
in the code, we unconditionally called fix_fields on the new
reference. A change happened in commit 4d0a79bf4e on branch
mysql-trunk-wl9603, when Evgen removed the call to fix_fields for the
newly created reference for some reason.

In our repro, fix field isn't called for it anywhere else either
before we get to the call of resolve_type mentioned above, hence the
ASSERT error.

[1] See added test case in window_functions_bugs.test

Solution:

Reintroduce the call to fix_fields, but now only conditionally, if
the referenced alias hasn't yet been resolved. (I tried to call in
unconditionally, but that caused another assert).

Regressions passed.

Change-Id: Ic882b92a4b101ec1c610458a9a77415cfd6c0fed